fix(crew): split server-function runtime boundary
