+++ apps/crew/src/routes/events.tsx
@@ -1,4 +1,9 @@
-import { createFileRoute, Link } from "@tanstack/react-router"
+import {
+  createFileRoute,
+  Link,
+  Outlet,
+  useRouterState,
+} from "@tanstack/react-router"
 import { formatCrewValue } from "@/lib/crew-event-display"
 import {
   calculateSetupProgress,
@@ -13,6 +18,13 @@ export const Route = createFileRoute("/events")({
 
 function EventsPage() {
   const { events } = Route.useLoaderData()
+  const isEventsIndex = useRouterState({
+    select: (state) => state.location.pathname.replace(/\/$/, "") === "/events",
+  })
+
+  if (!isEventsIndex) {
+    return <Outlet />
+  }
 
   return (
     <main className="mx-auto max-w-6xl space-y-6 px-4 py-8 sm:px-6">
