+++ apps/crew/src/server-fns/crew-event-settings-fns.ts
@@ -1,70 +1,26 @@
-import { createId } from "@paralleldrive/cuid2"
+// @lat: [[crew#Server Function Runtime Boundary]]
 import { createServerFn } from "@tanstack/react-start"
-import { desc, eq } from "drizzle-orm"
 import { z } from "zod"
-import { getDb } from "../db"
-import {
-  CREW_CONCIERGE_STATUS,
-  CREW_EVENT_LIFECYCLE,
-  CREW_PLAN,
-  type CrewEventSettings,
-  crewEventSettingsTable,
-} from "../db/schemas/crew-event-settings"
-import { type Competition, competitionsTable } from "../db/schemas/competitions"
-import { teamTable } from "../db/schemas/teams"
-import { requireLocalCrewOperatorAccess } from "../server/crew-local-access"
-import { generateSlug } from "../utils/slugify"
 
-type CrewEventCompetition = Pick<
-  Competition,
-  | "id"
-  | "organizingTeamId"
-  | "competitionTeamId"
-  | "slug"
-  | "name"
-  | "description"
-  | "startDate"
-  | "endDate"
-  | "timezone"
-  | "status"
-  | "visibility"
-  | "competitionType"
->
-
-export interface CrewEventDetails {
-  settings: CrewEventSettings
-  competition: CrewEventCompetition
-}
-
-function requireLocalCrewSettingsAccess() {
-  requireLocalCrewOperatorAccess("Crew event settings")
-}
+export type { CrewEventDetails } from "../server/crew-event-settings.server"
 
 const lifecycleSchema = z.enum([
-  CREW_EVENT_LIFECYCLE.DRAFT,
-  CREW_EVENT_LIFECYCLE.SETUP,
-  CREW_EVENT_LIFECYCLE.IMPORTING,
-  CREW_EVENT_LIFECYCLE.READY,
-  CREW_EVENT_LIFECYCLE.ARCHIVED,
+  "draft",
+  "setup",
+  "importing",
+  "ready",
+  "archived",
 ])
-
 const conciergeStatusSchema = z.enum([
-  CREW_CONCIERGE_STATUS.NOT_STARTED,
-  CREW_CONCIERGE_STATUS.IN_PROGRESS,
-  CREW_CONCIERGE_STATUS.READY,
-  CREW_CONCIERGE_STATUS.BLOCKED,
+  "not_started",
+  "in_progress",
+  "ready",
+  "blocked",
 ])
-
-const crewPlanSchema = z.enum([
-  CREW_PLAN.SELF_SERVE,
-  CREW_PLAN.CONCIERGE,
-  CREW_PLAN.FULL_PLATFORM,
-])
-
+const crewPlanSchema = z.enum(["self_serve", "concierge", "full_platform"])
 const dateStringSchema = z
   .string()
   .regex(/^\d{4}-\d{2}-\d{2}$/, "Date must be in YYYY-MM-DD format")
-
 const nullableTextInput = z
   .string()
   .trim()
@@ -88,7 +44,7 @@ const createCrewEventInputSchema = z.object({
   sourceEventUrl: nullableTextInput,
   externalRegistrationUrl: nullableTextInput,
   acquisitionSource: nullableTextInput,
-  crewPlan: crewPlanSchema.default(CREW_PLAN.SELF_SERVE),
+  crewPlan: crewPlanSchema.default("self_serve"),
   settings: nullableTextInput,
 })
 
@@ -98,7 +54,7 @@ const createCrewSettingsForCompetitionInputSchema = z.object({
   sourceEventUrl: nullableTextInput,
   externalRegistrationUrl: nullableTextInput,
   acquisitionSource: nullableTextInput,
-  crewPlan: crewPlanSchema.default(CREW_PLAN.SELF_SERVE),
+  crewPlan: crewPlanSchema.default("self_serve"),
   settings: nullableTextInput,
 })
 
@@ -116,93 +72,22 @@ const updateCrewEventSettingsInputSchema = z.object({
   settings: nullableTextInput,
 })
 
-const crewEventSelect = {
-  settings: crewEventSettingsTable,
-  competition: {
-    id: competitionsTable.id,
-    organizingTeamId: competitionsTable.organizingTeamId,
-    competitionTeamId: competitionsTable.competitionTeamId,
-    slug: competitionsTable.slug,
-    name: competitionsTable.name,
-    description: competitionsTable.description,
-    startDate: competitionsTable.startDate,
-    endDate: competitionsTable.endDate,
-    timezone: competitionsTable.timezone,
-    status: competitionsTable.status,
-    visibility: competitionsTable.visibility,
-    competitionType: competitionsTable.competitionType,
-  },
-}
-
-async function requireCrewEventCompetition(competitionId: string) {
-  const db = getDb()
-  const [competition] = await db
-    .select({
-      organizingTeamId: competitionsTable.organizingTeamId,
-    })
-    .from(competitionsTable)
-    .where(eq(competitionsTable.id, competitionId))
-    .limit(1)
-
-  if (!competition) {
-    throw new Error("Crew event not found")
-  }
-}
-
-async function requireOrganizingTeam(organizingTeamId: string) {
-  const db = getDb()
-  const [team] = await db
-    .select({ id: teamTable.id })
-    .from(teamTable)
-    .where(eq(teamTable.id, organizingTeamId))
-    .limit(1)
-
-  if (!team) {
-    throw new Error("Organizing team not found")
-  }
-}
-
-async function getCrewEventByCompetitionId(
-  competitionId: string,
-): Promise<CrewEventDetails | null> {
-  const db = getDb()
-  const [event] = await db
-    .select(crewEventSelect)
-    .from(crewEventSettingsTable)
-    .innerJoin(
-      competitionsTable,
-      eq(crewEventSettingsTable.competitionId, competitionsTable.id),
-    )
-    .where(eq(crewEventSettingsTable.competitionId, competitionId))
-    .limit(1)
-
-  return event ?? null
-}
-
 export const listCrewEventsFn = createServerFn({ method: "GET" }).handler(
-  async (): Promise<{ events: CrewEventDetails[] }> => {
-    requireLocalCrewSettingsAccess()
-
-    const db = getDb()
-    const events = await db
-      .select(crewEventSelect)
-      .from(crewEventSettingsTable)
-      .innerJoin(
-        competitionsTable,
-        eq(crewEventSettingsTable.competitionId, competitionsTable.id),
-      )
-      .orderBy(desc(crewEventSettingsTable.createdAt))
-
-    return { events }
+  async () => {
+    const { listCrewEvents } = await import(
+      "../server/crew-event-settings.server"
+    )
+    return listCrewEvents()
   },
 )
 
 export const getCrewEventFn = createServerFn({ method: "GET" })
   .inputValidator((data: unknown) => getCrewEventInputSchema.parse(data))
-  .handler(async ({ data }): Promise<{ event: CrewEventDetails | null }> => {
-    requireLocalCrewSettingsAccess()
-
-    return { event: await getCrewEventByCompetitionId(data.eventId) }
+  .handler(async ({ data }) => {
+    const { getCrewEvent } = await import(
+      "../server/crew-event-settings.server"
+    )
+    return getCrewEvent(data)
   })
 
 export const createCrewSettingsForCompetitionFn = createServerFn({
@@ -211,166 +96,29 @@ export const createCrewSettingsForCompetitionFn = createServerFn({
   .inputValidator((data: unknown) =>
     createCrewSettingsForCompetitionInputSchema.parse(data),
   )
-  .handler(async ({ data }): Promise<{ event: CrewEventDetails }> => {
-    requireLocalCrewSettingsAccess()
-    await requireCrewEventCompetition(data.competitionId)
-
-    const db = getDb()
-    const existing = await getCrewEventByCompetitionId(data.competitionId)
-    if (existing) return { event: existing }
-
-    await db.insert(crewEventSettingsTable).values({
-      competitionId: data.competitionId,
-      sourcePlatform: data.sourcePlatform,
-      sourceEventUrl: data.sourceEventUrl,
-      externalRegistrationUrl: data.externalRegistrationUrl,
-      acquisitionSource: data.acquisitionSource,
-      crewPlan: data.crewPlan,
-      settings: data.settings,
-    })
-
-    const event = await getCrewEventByCompetitionId(data.competitionId)
-    if (!event) {
-      throw new Error("Failed to create Crew event settings")
-    }
-
-    return { event }
+  .handler(async ({ data }) => {
+    const { createCrewSettingsForCompetition } = await import(
+      "../server/crew-event-settings.server"
+    )
+    return createCrewSettingsForCompetition(data)
   })
 
 export const createCrewEventFn = createServerFn({ method: "POST" })
   .inputValidator((data: unknown) => createCrewEventInputSchema.parse(data))
-  .handler(async ({ data }): Promise<{ event: CrewEventDetails }> => {
-    requireLocalCrewSettingsAccess()
-
-    const db = getDb()
-    const existingCompetition = await db
-      .select({ id: competitionsTable.id })
-      .from(competitionsTable)
-      .where(eq(competitionsTable.slug, data.slug))
-      .limit(1)
-
-    if (existingCompetition[0]) {
-      throw new Error(
-        "A competition with this slug already exists. Please choose a different slug.",
-      )
-    }
-
-    await requireOrganizingTeam(data.organizingTeamId)
-
-    let teamSlug = generateSlug(`${data.name}-event`)
-    let teamSlugIsUnique = false
-    let attempts = 0
-
-    while (!teamSlugIsUnique && attempts < 5) {
-      const existingTeam = await db
-        .select({ id: teamTable.id })
-        .from(teamTable)
-        .where(eq(teamTable.slug, teamSlug))
-        .limit(1)
-
-      if (!existingTeam[0]) {
-        teamSlugIsUnique = true
-      } else {
-        teamSlug = `${generateSlug(`${data.name}-event`)}-${createId().substring(0, 4)}`
-        attempts++
-      }
-    }
-
-    if (!teamSlugIsUnique) {
-      throw new Error("Could not generate unique slug for competition team")
-    }
-
-    const competitionTeamId = `team_${createId()}`
-    const competitionId = `comp_${createId()}`
-
-    await db.transaction(async (tx) => {
-      await tx.insert(teamTable).values({
-        id: competitionTeamId,
-        name: `${data.name} (Event)`,
-        slug: teamSlug,
-        type: "competition_event",
-        parentOrganizationId: data.organizingTeamId,
-        description: `Competition event team for ${data.name}`,
-        creditBalance: 0,
-      })
-
-      await tx.insert(competitionsTable).values({
-        id: competitionId,
-        organizingTeamId: data.organizingTeamId,
-        competitionTeamId,
-        name: data.name,
-        slug: data.slug,
-        description: data.description,
-        startDate: data.startDate,
-        endDate: data.endDate,
-        timezone: data.timezone,
-        competitionType: "in-person",
-      })
-
-      await tx.insert(crewEventSettingsTable).values({
-        competitionId,
-        sourcePlatform: data.sourcePlatform,
-        sourceEventUrl: data.sourceEventUrl,
-        externalRegistrationUrl: data.externalRegistrationUrl,
-        acquisitionSource: data.acquisitionSource,
-        crewPlan: data.crewPlan,
-        settings: data.settings,
-      })
-    })
-
-    const event = await getCrewEventByCompetitionId(competitionId)
-    if (!event) {
-      throw new Error("Failed to create Crew event")
-    }
-
-    return { event }
+  .handler(async ({ data }) => {
+    const { createCrewEvent } = await import(
+      "../server/crew-event-settings.server"
+    )
+    return createCrewEvent(data)
   })
 
 export const updateCrewEventSettingsFn = createServerFn({ method: "POST" })
   .inputValidator((data: unknown) =>
     updateCrewEventSettingsInputSchema.parse(data),
   )
-  .handler(async ({ data }): Promise<{ event: CrewEventDetails }> => {
-    requireLocalCrewSettingsAccess()
-    await requireCrewEventCompetition(data.competitionId)
-
-    const updateData: Partial<typeof crewEventSettingsTable.$inferInsert> = {
-      updatedAt: new Date(),
-    }
-
-    if (data.crewOnly !== undefined) updateData.crewOnly = data.crewOnly
-    if (data.sourcePlatform !== undefined) {
-      updateData.sourcePlatform = data.sourcePlatform
-    }
-    if (data.sourceEventUrl !== undefined) {
-      updateData.sourceEventUrl = data.sourceEventUrl
-    }
-    if (data.externalRegistrationUrl !== undefined) {
-      updateData.externalRegistrationUrl = data.externalRegistrationUrl
-    }
-    if (data.lifecycle !== undefined) updateData.lifecycle = data.lifecycle
-    if (data.conciergeStatus !== undefined) {
-      updateData.conciergeStatus = data.conciergeStatus
-    }
-    if (data.crewPlan !== undefined) updateData.crewPlan = data.crewPlan
-    if (data.fullPlatformCreditCents !== undefined) {
-      updateData.fullPlatformCreditCents = data.fullPlatformCreditCents
-    }
-    if (data.acquisitionSource !== undefined) {
-      updateData.acquisitionSource = data.acquisitionSource
-    }
-    if (data.settings !== undefined) updateData.settings = data.settings
-
-    const db = getDb()
-    await db
-      .update(crewEventSettingsTable)
-      .set(updateData)
-      .where(eq(crewEventSettingsTable.competitionId, data.competitionId))
-
-    const event = await getCrewEventByCompetitionId(data.competitionId)
-    if (!event) {
-      throw new Error("Crew event not found")
-    }
-
-    return { event }
+  .handler(async ({ data }) => {
+    const { updateCrewEventSettings } = await import(
+      "../server/crew-event-settings.server"
+    )
+    return updateCrewEventSettings(data)
   })
